requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
pdfplumber
pytesseract
pillow